health-monitor-app/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
matplotlib==3.7.2
psutil==5.9.5
